+++ public/src/components/UserList.js
@@ -2,19 +2,16 @@
 // const fs = require('fs');
  
 import React, {Component, Fragment} from 'react';
-import USERS from '../MOCK_DATA.js'
+import USERS from '../scripts/MOCK_DATA.js'
 
-
-
-console.log('users', USERS);
 class USERLIST extends Component {
   render() {
    return <Fragment>
     <h1>Users</h1>
     <p>{USERS.length} users</p>
     {USERS.map(user => {
-      return <div>
-        ({user.first_name})
+      return <div key={user.id}>
+        {user.originalId} {user.first_name}  {user.last_name} 
         </div>
     })}
     </Fragment>
