+++ scripts/seed.js
@@ -1,0 +1,1019 @@
+let USERS =`1,Elfreda,Muirhead
+2,Nonnah,Bleiman
+3,Bryn,Fenton
+4,Susan,Blackman
+5,Lorianna,Lenaghen
+6,Annice,Ivanenkov
+7,Marietta,Tomasello
+8,Andrea,Peasey
+9,Dermot,Gluyus
+10,Georgianna,Ohm
+11,Riccardo,Scrase
+12,Cassi,Barthot
+13,Armin,Belsham
+14,Alyce,Cord
+15,Breanne,Goodbairn
+16,Andros,Jackways
+17,Sterling,Whiston
+18,Deny,Waycott
+19,Donavon,Burrill
+20,Fairleigh,Pavlov
+21,Yalonda,Ballefant
+22,Shelby,Duligall
+23,Georgine,Caccavari
+24,Thaxter,Hanlon
+25,Drucill,Livezley
+26,Nicholle,Frend
+27,Dorri,Longmire
+28,Hadlee,Harrowing
+29,Toby,Caltera
+30,Charmian,Worge
+31,Halimeda,Andriulis
+32,Conni,Caghy
+33,Dalia,Braz
+34,Davine,Olenchikov
+35,Ardyth,Moralis
+36,Ganny,Peirpoint
+37,Steffi,Hanmore
+38,Terry,Fey
+39,Reynold,Cahalan
+40,Rosabel,Maunders
+41,Bree,McVeighty
+42,Franni,MacRedmond
+43,Keir,Longhorne
+44,Marrilee,Viste
+45,Louis,Deadman
+46,Emili,Patroni
+47,Nikki,MacKaig
+48,Theodora,Greystoke
+49,Jacquelyn,Drillingcourt
+50,Aimee,Dyball
+51,Ransell,Hellier
+52,Krystalle,Wannan
+53,Torr,Girdlestone
+54,Kahaleel,Ovanesian
+55,Jayne,Elsay
+56,Tabbie,Rennles
+57,Arie,Mara
+58,Julissa,Luesley
+59,Benoite,Charnick
+60,Vaclav,Plumptre
+61,Adolf,Sara
+62,Benjamen,Lodevick
+63,Amye,Bantham
+64,Flora,Blare
+65,Fawne,Gannon
+66,Caren,Borland
+67,Kerri,Stenson
+68,Blair,Barnhart
+69,Eadith,Meade
+70,Iris,Rennocks
+71,Dennie,Dreschler
+72,Chaunce,Broker
+73,Roderick,Gregor
+74,Mattheus,Joplin
+75,Phylys,Haselhurst
+76,Vernor,Emmines
+77,Robbyn,Morando
+78,Ado,Perchard
+79,Lewiss,Hryncewicz
+80,Gael,Bosley
+81,Renae,Brger
+82,Dacey,Ochiltree
+83,Juli,Toderi
+84,Townie,Allsupp
+85,Rea,Romagnosi
+86,Tomasina,Jacquest
+87,Demetri,Stiebler
+88,Trumaine,Gibbon
+89,Halli,Symson
+90,Josias,Portal
+91,Pauli,Tregea
+92,Griffie,Meardon
+93,Demetri,Dacey
+94,Jeri,MacTerlagh
+95,Debee,Glander
+96,Sherie,Shellard
+97,Noami,Gregr
+98,Sadella,Bugg
+99,Bernardina,McKeighen
+100,Kathi,Scandred
+101,Robbie,Looker
+102,Jeana,Sprasen
+103,Torrence,Ronaghan
+104,Aeriel,Erik
+105,Janeczka,Shutte
+106,Ingmar,Minshall
+107,Farah,Hurburt
+108,Melony,Gasticke
+109,Maisie,Stedell
+110,Clea,Fernely
+111,Jere,Wyllie
+112,Kurtis,Zamora
+113,Toma,Stopps
+114,Myra,Dussy
+115,Claybourne,Grandison
+116,Lexi,Gorger
+117,Betti,Redferne
+118,Karilynn,Coen
+119,Emmye,Crate
+120,Gretchen,Trevino
+121,Carling,Drakes
+122,Frederic,Nield
+123,Claudio,Raspel
+124,Hillary,Aldhous
+125,Ricca,Elix
+126,Lyon,Bozward
+127,Darelle,Norville
+128,Eula,Shutt
+129,Cora,Meo
+130,Harrie,Capitano
+131,Erika,Grichukhin
+132,Lucienne,Housaman
+133,Suellen,Boldock
+134,Cart,Lammerts
+135,Keri,Gibb
+136,Viviyan,MacNeilly
+137,Claudia,Stork
+138,Berny,Tackes
+139,Skyler,Storrs
+140,Mose,Horley
+141,Cordelia,Cawcutt
+142,Vanessa,Kiss
+143,Townsend,Tebbut
+144,Elfra,Pinchback
+145,Manda,Va
+146,Glendon,Danson
+147,Elysee,Hyndman
+148,Sigfr,Belone
+149,Arlette,Flag
+150,Austen,Hare
+151,Carry,Couldwell
+152,Peder,Tammadge
+153,Gonzales,McCloughlin
+154,Karoly,Livingstone
+155,Elayne,Simnor
+156,Tilly,Versey
+157,Boigie,Praill
+158,Tracee,Meech
+159,Adolpho,Jurisic
+160,Neile,Sutheran
+161,Meredithe,Rubinshtein
+162,Pierre,Crissil
+163,Phara,Dagnan
+164,Angel,Scampion
+165,Joycelin,Basten
+166,Ilaire,Pyatt
+167,Jamal,McGettrick
+168,Basia,Dickey
+169,Skyler,O'Kuddyhy
+170,Gunar,Claw
+171,Burtie,Paddon
+172,Merrie,Cobbald
+173,Ninon,Mylechreest
+174,Jared,Seares
+175,Borden,Shulem
+176,Stanfield,Kelle
+177,Keri,Winder
+178,Dorian,Butler
+179,Jillana,Weaver
+180,Rand,Adolphine
+181,Pattin,Schubuser
+182,Rab,Shefton
+183,Quentin,Emblen
+184,Enrika,Ancliffe
+185,Bertha,Ripon
+186,Tamma,Ibbetson
+187,Stavros,Berry
+188,Maximilian,Gorusso
+189,Lia,Oak
+190,Lorry,Halson
+191,Ezekiel,Meeking
+192,Hedvige,Vedenichev
+193,Amalita,Janway
+194,Theresa,Gimeno
+195,Tulley,Cholerton
+196,Alisun,Loreit
+197,Frazer,Briskey
+198,Georgia,Everist
+199,Ginni,Ansell
+200,Hynda,Waterhous
+201,Vikky,Assur
+202,Felipa,Haslock
+203,Ber,Orrell
+204,Selena,Lythgoe
+205,Sigismundo,Scarse
+206,Tallie,Wakeling
+207,Giavani,Kadd
+208,Giustino,Etock
+209,Elga,Ranger
+210,Bernardine,MacRierie
+211,Shannan,Pavlenkov
+212,Morse,Seabert
+213,Ermin,Denslow
+214,Janek,Prichet
+215,Rebecka,Haddeston
+216,Kamillah,Ferraresi
+217,Vanni,Sheerin
+218,Steve,Vlasov
+219,Marshall,Buddell
+220,Bernadina,Soal
+221,Annadiane,Tripon
+222,Eartha,Coulton
+223,Constantino,Norrington
+224,Dorey,Swindley
+225,Jamison,Olek
+226,Kasper,enden
+227,Stirling,Sanderson
+228,Liana,Beesey
+229,Alastair,Leverette
+230,Boy,Silwood
+231,Madalyn,De Andreis
+232,Flss,Antill
+233,Wally,Congram
+234,Chauncey,Ritmeyer
+235,Twyla,Winscomb
+236,Orlando,Jameson
+237,Orlan,Biaggioni
+238,Vergil,Chooter
+239,Danya,Happs
+240,Robin,Beaney
+241,Berke,McClo
+242,Shani,Chessman
+243,Boyd,Fulker
+244,Feliza,Brimson
+245,Camala,Paterson
+246,Robinette,Simonato
+247,Truman,Cousans
+248,Courtnay,Enocksson
+249,Carey,Gricks
+250,Filip,Ellam
+251,Yolanda,Gossage
+252,Re,Saunton
+253,Riley,De Bell
+254,Oona,Wymer
+255,Isak,Dighton
+256,Bria,MacTrustam
+257,Hashim,Chazelle
+258,Lester,Engeham
+259,Fawne,Cleiment
+260,Orella,Cotilard
+261,Benn,Pelling
+262,Laura,Duchan
+263,Sayers,Guyver
+264,Hedi,Dredge
+265,Inglis,Darwen
+266,Ginnifer,Weigh
+267,Morty,Strognell
+268,Nelie,Hucke
+269,Joaquin,Lethlay
+270,Barnard,Braban
+271,Conn,Shemming
+272,Nickolas,Besnardeau
+273,Che,Bartleet
+274,Bard,Marzella
+275,Venus,Footer
+276,Ilario,Fullylove
+277,Ermentrude,Tender
+278,Laverne,Harkin
+279,Ermengarde,Cordeiro
+280,Lolita,Rosenshine
+281,Case,Grealey
+282,Brear,Neillans
+283,Evelin,Wynrehame
+284,Linell,Doni
+285,Orsa,Sperwell
+286,Alfy,Eskriett
+287,Susanetta,Ellick
+288,Elliott,Raffeorty
+289,Glyn,Prinett
+290,Hercule,Staning
+291,Avis,Janatka
+292,Raine,Dray
+293,Booth,Dahlen
+294,Lucilia,Durrance
+295,Garnet,Kimble
+296,Riki,Lemme
+297,Ashley,Grinston
+298,Gray,Fairweather
+299,Brinna,Clackson
+300,Averill,Sweeten
+301,Gilburt,Hawkeswood
+302,Viviene,Malkie
+303,Betta,Zamboniari
+304,William,Bruun
+305,Katrine,Rash
+306,Jeannette,Craighall
+307,Gilberta,Nockells
+308,Lorianne,Bamforth
+309,Barbey,Bentsen
+310,Aurelia,Grunwald
+311,Allyn,Godbald
+312,Theodora,Bogaert
+313,Vassily,Broseke
+314,Ellynn,Pimblotte
+315,Hyacinthia,Trevaskus
+316,Ediva,Kitchingham
+317,Dominique,Moulin
+318,Katinka,Kniveton
+319,Cordell,Cardow
+320,Clemens,Gwinnel
+321,Cleo,Clewes
+322,Dacy,Courtier
+323,Amabelle,Gentric
+324,Leanora,de Quincey
+325,Leonora,O'Sculley
+326,Herrick,Legge
+327,Josefa,Vittori
+328,Theo,Taleworth
+329,Vanessa,Huxton
+330,Norby,Briar
+331,Kippie,Schulze
+332,Dedie,Greave
+333,Christi,Le Floch
+334,Wilmar,Kivlehan
+335,Serena,Unsworth
+336,Guinna,Blackett
+337,Moira,Fayer
+338,Dasie,Rhubottom
+339,Dory,Dunlap
+340,Jacquenetta,Baldick
+341,Nalani,Di Carlo
+342,Steward,Kenefick
+343,Wayne,Margarson
+344,Joleen,Bogie
+345,Clementine,Wilmot
+346,Millisent,Feldhorn
+347,Iris,Quick
+348,Tore,Lazenbury
+349,Merrilee,Anneslie
+350,Amil,Hawkrge
+351,Carmelia,Artz
+352,Elnora,Shapcott
+353,Snee,Ayrs
+354,Lotte,Vaughton
+355,Sonnnie,Silversm
+356,Brig,Woodworth
+357,Hallsy,Crunkhorn
+358,Tuesday,Yuryaev
+359,Ollie,Cave
+360,Carmina,Tunna
+361,Vladimir,Rubinowitsch
+362,Madel,O'Kelly
+363,Cookie,Monkton
+364,Edward,Pauletto
+365,Riobard,MacCarter
+366,Morton,Fuzzens
+367,Elva,Philcock
+368,Mitchel,Grumell
+369,Will,Danett
+370,Ingeborg,Erasmus
+371,Vally,Yarnall
+372,Kristan,Gres
+373,Ingaberg,Giacoppo
+374,Reta,Taverner
+375,Mario,Reinger
+376,Tamma,Scarisbrick
+377,Yasmin,Charlin
+378,Herculie,Dunnaway
+379,Joleen,Comar
+380,Caitrin,Tinson
+381,Tomas,Aldgate
+382,Maurits,Mirralls
+383,Victoria,Davoren
+384,Gabby,Ference
+385,Gabriellia,Haking
+386,Drugi,Scouller
+387,Dorothy,Wooster
+388,Oriana,Sarson
+389,Selie,Argabrite
+390,Gerome,Buggs
+391,Lucine,Guerola
+392,Leonardo,Lutton
+393,Anna-diane,Hearty
+394,Janie,Barthelme
+395,Goran,Sutherley
+396,Jemima,Hillburn
+397,Benedikta,Wooler
+398,Alford,Elwin
+399,Forrest,Jewel
+400,Karie,Wainscot
+401,Jacquette,Ciobotaru
+402,Filmer,Seamen
+403,Therese,Beare
+404,Maddie,Laven
+405,Happy,Fishenden
+406,Denny,Wooton
+407,Luigi,Aveyard
+408,Marni,Baudouin
+409,Byran,Brashaw
+410,Jacinthe,Muckian
+411,Barnett,Scutching
+412,Arline,Kordt
+413,Ewell,Massimo
+414,Codie,Burnel
+415,Herve,Pirolini
+416,Brande,Macci
+417,Camilla,Faier
+418,Merci,Briggs
+419,Phelia,Sprey
+420,Benjie,Coarser
+421,Murial,Massen
+422,Urson,Vaux
+423,Arlyn,Curds
+424,Donella,Berzin
+425,Torey,Folder
+426,Ali,Domb
+427,Hynda,Wordington
+428,Suzette,Gatherell
+429,Legra,Gerrish
+430,Willie,Wickerson
+431,Ingeborg,Benoey
+432,Libbi,Standall
+433,Cam,Cornwall
+434,Mair,O'Fergus
+435,Beatrix,Warcup
+436,Mar,Dayley
+437,Onfroi,Prest
+438,Grantley,Legan
+439,Karee,Clemmen
+440,Wyatan,Nightingale
+441,Meaghan,Breeder
+442,Alfi,Node
+443,Ashleigh,Meachem
+444,Alma,Haddacks
+445,Allyson,Tomek
+446,Kimberley,Nesey
+447,Eileen,Dooher
+448,Brigitta,Gai
+449,Ingeborg,Pedler
+450,Winston,Castelli
+451,Hunter,Cohalan
+452,Alvira,Booton
+453,Rea,Felix
+454,Jojo,Dahlberg
+455,Winne,Errichi
+456,Hillary,MacCafferky
+457,Al,Rounsivall
+458,Dayle,Dict
+459,Cordy,Wrathmell
+460,Melicent,Smalcombe
+461,Dell,Morican
+462,Allister,Shanahan
+463,Drusie,Downes
+464,Duane,Tomek
+465,Ginevra,Saberton
+466,Haleigh,Lafaye
+467,Sonni,McHugh
+468,Trina,Gosdin
+469,Jere,Elstob
+470,Dacia,Maytom
+471,Clovis,Ackroyd
+472,Kennedy,Henbury
+473,Dasha,Mitchall
+474,Carlo,McGilmartin
+475,Catharina,Duquesnay
+476,Marillin,Izkovitch
+477,Nerti,Shoulders
+478,Hector,Desbrow
+479,Korey,Furbank
+480,Deb,Amesbury
+481,Allyn,Wyldes
+482,Far,Brownscombe
+483,Petra,McDa
+484,Deb,Brasner
+485,Latashia,Tippell
+486,Ernesta,Littler
+487,Miranda,Boutcher
+488,Aindrea,Slogrove
+489,Morten,Rushe
+490,Fan,Ellul
+491,Gregorio,McCart
+492,Jacobo,Jarlmann
+493,Bdy,Merrilees
+494,Kile,Saint
+495,Alie,Hapke
+496,Dorena,Purvey
+497,Mathilda,Pughsley
+498,Berget,Binding
+499,Ebeneser,Kenworthey
+500,Deonne,Ruxton
+501,Breanne,Goodread
+502,Verla,MacPake
+503,Pete,Inold
+504,Brande,Carlson
+505,Nestor,Cossentine
+506,Thane,Ledwith
+507,Pembroke,Dalinder
+508,Marcellus,Piggrem
+509,Leanna,Vice
+510,Elsey,Spera
+511,Claudina,Kemsley
+512,Cesar,McLelland
+513,Shane,Muddicliffe
+514,Leshia,Grigoliis
+515,Werner,O' Hogan
+516,Ethelda,Pontain
+517,Emelia,Simoncini
+518,Kathrine,Hoyles
+519,Averil,Jeffress
+520,Smith,Coppard
+521,Annis,Mitchley
+522,Cooper,Demoge
+523,Melisande,Handes
+524,Elliott,Haville
+525,Ransell,Malter
+526,Eran,Beynon
+527,Hugo,Hender
+528,Damita,Disbrow
+529,Lelah,McIlreavy
+530,Margarita,Osbston
+531,Errol,Chartres
+532,Evan,Scoterbosh
+533,Catha,Espinoza
+534,Bennett,Dower
+535,Forster,Nisbith
+536,Valli,Saura
+537,Becky,Assad
+538,Meredithe,Feragh
+539,Kip,Esplin
+540,Shelby,Thirwell
+541,Prisca,Dumbelton
+542,Jacobo,Sturrock
+543,Carlyn,Frame
+544,Normy,St Ledger
+545,Traci,Zink
+546,Adore,Glusby
+547,Trueman,Scourge
+548,Jodie,Hanniger
+549,Starlin,Basham
+550,Jodi,Le Pruvost
+551,Mac,Langforth
+552,Sayers,Karolewski
+553,Seline,Clinch
+554,Flossy,Billiard
+555,Clementine,Bream
+556,Margaretta,Spencers
+557,Almeta,Rouchy
+558,Bradly,Miall
+559,Abigael,Bonsul
+560,Terencio,Charrette
+561,Craggie,Daniell
+562,Levi,Gehrtz
+563,Celestyn,Lukas
+564,Daisey,Beringer
+565,Hilario,Pawlik
+566,Peterus,Westmoreland
+567,Gianina,Raiker
+568,Sansone,Duplock
+569,Jarret,Mole
+570,Claudio,Orsi
+571,Elysee,Paolacci
+572,Rozalin,Keepe
+573,Garret,Frayn
+574,Udall,Wiszniewski
+575,Duff,Halgarth
+576,Regen,Sergant
+577,Philippine,Aubert
+578,Ines,Grimmolby
+579,Torrin,Grinnov
+580,Michaeline,Eubank
+581,Antonina,Pyburn
+582,Mark,Bruckman
+583,Corabelle,Elliot
+584,Reilly,Reedie
+585,Linda,Vasyutin
+586,Terza,McLorinan
+587,Ana,Orehead
+588,Harmony,Blackborn
+589,Sig,Brannan
+590,Amye,Renzo
+591,Rochester,Tittershill
+592,Charlton,Tapping
+593,Lazare,Bangham
+594,Jessika,Reith
+595,Jdavie,Rodell
+596,Zena,Rubinov
+597,Schuyler,Prestie
+598,Tommi,Tolmie
+599,Yale,Pfeiffer
+600,Alaster,Shapland
+601,Eberto,Rehorek
+602,Marlyn,Tomaszynski
+603,Arlyn,Lyttle
+604,Collie,Merrisson
+605,Brade,Larver
+606,Keir,Mewitt
+607,Jaime,Cleal
+608,Maurits,Ruegg
+609,Elijah,Heavyse
+610,Sharla,Fazakerley
+611,Colin,Breache
+612,Berty,Ricciardiello
+613,Addison,Bexley
+614,Helen,Bavin
+615,Ellissa,Ullyott
+616,Dorris,Jacobowits
+617,Remington,Follin
+618,Tabbatha,Ledwge
+619,Reuben,Howlings
+620,Laurianne,Castana
+621,Ariel,Baddow
+622,Bard,Stormonth
+623,Jarr,Sutherns
+624,Leanor,Fretson
+625,Pieter,Tottie
+626,Guillema,Overshott
+627,Starla,Ducrow
+628,Roseline,Aler
+629,Deni,Morewood
+630,Clerkclaude,Truscott
+631,Spense,Cardus
+632,Ulrike,Pascoe
+633,Dwayne,McCleverty
+634,Flossie,Ternott
+635,Quintilla,Easey
+636,Johannah,Ralston
+637,Maddie,Gerald
+638,Martha,Atling
+639,Christean,Balthasar
+640,Bellina,Colrein
+641,Jean,Fdyment
+642,Siffre,Preedy
+643,Jeffrey,Bawles
+644,Andeee,Grier
+645,Justus,Kynson
+646,Phineas,Wildor
+647,Cindi,Abelevitz
+648,Benoite,Arnke
+649,Levey,Ablott
+650,Carroll,Stanistrete
+651,Blinny,Yeats
+652,Clyde,Elham
+653,Humberto,Luxmoore
+654,Clo,Catcherse
+655,Caria,Hourigan
+656,Shep,Ehlerding
+657,Zita,Stut
+658,Monah,Gersam
+659,Georgie,Jiricka
+660,Salvatore,Methley
+661,Britte,Clute
+662,Fransisco,Maryet
+663,Lurline,Emmatt
+664,Jonell,Crowder
+665,Remington,Fenny
+666,Sigismund,Swinnard
+667,Aloysia,Spinke
+668,Davie,Karpfen
+669,Lavina,Gawkroge
+670,Lynnet,Caws
+671,Darin,Khadir
+672,Sorcha,Parfrey
+673,Aloisia,Croasdale
+674,Nathanil,Orriss
+675,Floyd,Grigoli
+676,Jose,Syseland
+677,Vera,Rosenhaupt
+678,Pincas,Hawkswood
+679,Shirleen,Padwick
+680,Logan,Jankiewicz
+681,Robyn,Raggles
+682,Jandy,Argile
+683,Katlin,Lightwing
+684,Kirbie,Pimme
+685,Serge,Scamal
+686,Liliane,Rosenauer
+687,Ker,Cuell
+688,Bettina,Mawson
+689,Michell,Chern
+690,Goldina,MacRury
+691,Currie,Furmage
+692,Loutitia,Flett
+693,Bobbe,Vieyra
+694,Carolyn,Bourke
+695,Madlin,Ranscome
+696,Nikola,Gauge
+697,Nikolos,Stuchburie
+698,Gerrie,Sweeny
+699,Herb,Aitcheson
+700,Justinn,Tooley
+701,Yvette,Deniskevich
+702,Eleanor,Ferrari
+703,Roxi,Quilter
+704,Emmit,Johann
+705,Parnell,Astin
+706,Antonietta,Pattullo
+707,Ruperto,Luggar
+708,Ezri,Caldero
+709,Aloin,Sanney
+710,Barbi,Lindenboim
+711,Beatrisa,Jacketts
+712,Rosamund,Paulet
+713,Bethina,Woodroffe
+714,Sam,Kieff
+715,Debi,Marrow
+716,Nell,Willavoys
+717,Meghann,Braen
+718,Matthiew,Saffe
+719,Blane,Muzzlewhite
+720,Claudie,Hruska
+721,Brietta,Strewthers
+722,Haleigh,Tattersdill
+723,Corly,Dytham
+724,Sharleen,Yonnie
+725,Isabelita,Fahrenbacher
+726,Carmine,Laird
+727,Rory,Clissold
+728,Casey,Estoile
+729,Allyce,Micallef
+730,Christophorus,Rhucroft
+731,Kendal,Bonar
+732,Orel,Ollive
+733,Carlita,Hews
+734,Corny,Hearns
+735,Georgianne,Judd
+736,Dorita,Pugh
+737,Marla,Furst
+738,Devland,Kinzett
+739,Wally,McBre
+740,Neron,Penticost
+741,Terrell,Reie
+742,Traci,Letterick
+743,Joellen,Briggdale
+744,Giffer,Schultes
+745,Cecilia,Iacovaccio
+746,Carie,Underwood
+747,Moria,Oldmeadow
+748,Robin,Snedden
+749,Logan,Drains
+750,Molly,Attfield
+751,Bennie,Pickin
+752,Sibel,Borrington
+753,Mariellen,Ungerer
+754,Marshall,Reddell
+755,Gertie,Fairbrge
+756,Chrotoem,Belchamber
+757,Aimee,Filov
+758,Calv,McCreery
+759,Mollee,Stennett
+760,Vladamir,Wilmot
+761,Vanny,Gibson
+762,Sara-ann,Bolton
+763,Zaneta,Dahler
+764,Orin,Barthrup
+765,Pooh,Cottrell
+766,Talia,Overil
+767,Julienne,Victor
+768,Bail,Yare
+769,Kevina,Huddles
+770,Kynthia,Tire
+771,Sherlock,Gambell
+772,Nettie,Chaffyn
+773,Fran,Allenby
+774,Kimberlee,Cossins
+775,Marilyn,Himsworth
+776,Derrick,Fell
+777,Oona,Wedon
+778,Luce,Buckthought
+779,Monika,Loxston
+780,Ami,Dearman
+781,Erinna,Quarrell
+782,Gloria,Nesbit
+783,Alleyn,Hallett
+784,Muffin,Rains
+785,Fabiano,Tatem
+786,Chiarra,McNiven
+787,Ardath,Siberry
+788,Brocky,Shimmans
+789,Sukey,Furley
+790,Godwin,Dwight
+791,Kennan,Stelfax
+792,Jayme,Newarte
+793,Penn,Le Grand
+794,Thomasin,Brawley
+795,Quincey,Scyone
+796,Ninon,Dumigan
+797,Louise,Espada
+798,Kerrin,Clemas
+799,Alana,McGillivrie
+800,Tremain,Creswell
+801,Nanete,Mdleweek
+802,Deva,Oxburgh
+803,Randene,Ransfield
+804,Roderic,Wilkennson
+805,Onfre,Sandom
+806,Tiena,Playhill
+807,Morgan,Aglione
+808,Kimbell,Milnes
+809,Agace,Ellaway
+810,Zechariah,Amthor
+811,Johan,Wyre
+812,Ker,Garnett
+813,Cullan,Jiri
+814,Jaye,Gilbee
+815,Kaela,Zoephel
+816,Victoir,Laneham
+817,Vanna,Jory
+818,Oriana,Pancoust
+819,Jeana,McTeague
+820,Diandra,Fishpool
+821,Joell,McWhinnie
+822,Pembroke,Lightwing
+823,Tully,Gottelier
+824,Skippie,Rossbrook
+825,Tiphany,Sarvar
+826,Hinda,Kinde
+827,Silvio,Assaf
+828,Milt,Hemphill
+829,Ali,Decent
+830,Shaughn,Cristofanini
+831,Mata,Dewane
+832,Patrica,Maghull
+833,Babs,Waterson
+834,Jannel,Tubby
+835,Anselm,Evequot
+836,Leighton,Mewrcik
+837,Lorry,Glascott
+838,Rosemarie,Powelee
+839,Berkie,Andriulis
+840,Joceline,Nicklen
+841,Dorie,Atger
+842,Mareah,Trevethan
+843,Winnie,Huonic
+844,Chaunce,Grimsdyke
+845,Fairfax,De Goey
+846,Austine,Chuter
+847,Harper,Seel
+848,Ranna,Livermore
+849,Penny,Anear
+850,En,Howes
+851,Drusie,Fudge
+852,Eberhard,Goodre
+853,Doro,Macek
+854,Ninette,Maro
+855,Nealson,L'Homme
+856,Jada,Blaasch
+857,Lem,Khosa
+858,Nativad,Penke
+859,Roana,McKdin
+860,Christy,Farrey
+861,Fredrick,Playhill
+862,Gerri,Vlasin
+863,Betteann,Rimell
+864,Libby,Haste
+865,Naomi,Livesley
+866,Meta,Cottham
+867,Rosabel,Botting
+868,Elliot,Redmille
+869,Freeman,Hazelgrove
+870,Damara,Priestner
+871,Elnore,Sedgeman
+872,Isacco,Bugdall
+873,Prissie,Gomar
+874,Andonis,Sherwyn
+875,Farrel,Edrge
+876,Mallorie,Syred
+877,Nicoli,Fountain
+878,Eimile,Morville
+879,Mareah,Renahan
+880,Dene,Niven
+881,Cathee,Dimitresco
+882,Allistir,Clavering
+883,Adriane,Cleworth
+884,Prentiss,Abbatt
+885,Sheela,Rubinovitch
+886,Michaella,Rossbrooke
+887,Carmine,Shilstone
+888,Filippa,Mustin
+889,Trevar,Vowells
+890,Dulcia,Clough
+891,Lennie,Mirralls
+892,Octavia,Siggs
+893,Amanda,Kubicka
+894,Win,Ratnege
+895,Ardella,Sich
+896,Katherine,Terzi
+897,Naoma,MacAindreis
+898,Billie,Draysey
+899,Ofilia,Priscott
+900,Cesaro,Letcher
+901,Jerri,Cristoferi
+902,Zeb,Petchey
+903,Arney,Pheazey
+904,Levon,Wiggett
+905,Livvyy,Sealeaf
+906,Fleming,Woltering
+907,Neda,Kedie
+908,Devina,Pirnie
+909,Aigneis,Lamberto
+910,Tymothy,Korneev
+911,Karla,Crasswell
+912,Devon,Fanshaw
+913,Sissy,Moubray
+914,Filmer,Brumfitt
+915,Noell,Possa
+916,Pamelina,Ander
+917,Chalmers,Le Barr
+918,Duke,Roughsedge
+919,Morgen,Prendeville
+920,Sharl,Bangs
+921,Daloris,Gazey
+922,Vachel,Nixon
+923,Chera,Lockner
+924,Chrystel,Dunmore
+925,Kira,Mincher
+926,Jeniece,Copnall
+927,Wit,Deignan
+928,Rubia,Balchen
+929,Evanne,Descroix
+930,Emery,Gunby
+931,La verne,Cromblehome
+932,Maurits,Hekkert
+933,Emmie,Finden
+934,Lauraine,Pilley
+935,Benjy,Marchand
+936,Law,Meech
+937,Cissy,Feilden
+938,Chris,Marlor
+939,Calypso,Record
+940,Roma,Brinkman
+941,Letisha,Laurand
+942,Farrel,Fearney
+943,Alberik,Kobieriecki
+944,Ragnar,Lambotin
+945,Ingrim,Bonds
+946,Chastity,Fowle
+947,Sioux,Alliston
+948,Judah,Itzkowicz
+949,Daffie,Pimlott
+950,Karee,Jouannin
+951,Norri,Vivyan
+952,Gustave,MacNeilly
+953,Nonie,Rushworth
+954,Babbie,Maccari
+955,Dal,Domoney
+956,Libbie,Sprigin
+957,Brockie,Lamar
+958,Malina,Coram
+959,Leslie,Leschelle
+960,Roy,Kapiloff
+961,Maxie,Geibel
+962,Osbourn,Shoobrge
+963,Ryon,Mephan
+964,Bradly,Finder
+965,Robbi,Maiklem
+966,Lennard,St Ledger
+967,Car,Landre
+968,Kermy,Gurley
+969,Ansell,Burril
+970,Harlen,Fenich
+971,Manuel,Cann
+972,Egon,Huckfield
+973,Legra,MacGillacolm
+974,Muire,Sarjent
+975,Alla,Ashforth
+976,Amata,Greet
+977,Tallia,Clemens
+978,Booth,Getcliff
+979,Ciel,Cassy
+980,Nola,Jankin
+981,Valera,Maddigan
+982,Yolande,Paton
+983,Cristy,Gowthorpe
+984,Donica,Tabourel
+985,Carey,Rycraft
+986,Kahlil,Pahlsson
+987,Ardisj,Watsam
+988,Carma,Berick
+989,Gorden,Faill
+990,Randene,Havvock
+991,Percy,Dedmam
+992,Latrina,Tresvina
+993,Gerek,Headingham
+994,Leroi,Edison
+995,Hyacinthie,Blockley
+996,Adoree,MacGillespie
+997,Nisse,Grellis
+998,Todd,Vedeniktov
+999,Ardyce,Macallam
+1000,Michail,Clemendet}`;
+
+USERS = USERS.split('\n');
+USERS = USERS.map(user => {
+  let cells = user.split(',');
+  return {
+    originalId: parseInt(cells[0]),
+    first_name: cells[1],
+    last_name: cells[2],
+  };
+});
+
+require('dotenv').config();
+const User = require('../models/userModel.js');
+const mongoose = require('mongoose');
+mongoose.connect(process.env.MONGODB_URI);
+
+USERS.forEach(user =>{
+  User.create(user);
+});
